+++ tests/main.rs
@@ -1,0 +1,2 @@
+mod lthash16;
+mod lthash32;